+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0 (2016-06-09)
+
+Added: Support for `!important` on either the font size or line height.
+Changed: Splits on a forward-slash (`/`) instead of a space (` `).
+Updated: Tests
+
 ## 1.0.1 (2015-09-29)
 
 Updated: Value parsing